requirements.txt
discord.py==2.3.0
praw==7.7.0
deep_translator==1.10.0
lxml==4.9.2
aiosqlite==0.19.0
beautifulsoup4==4.11.2
python-dotenv==1.0.0
sqlite3==2.4.1
